Create a `ResolvedVariantResult` instance once only for each variant.
