+++ subprojects/dependency-management/src/main/java/org/gradle/internal/component/model/AbstractComponentGraphResolveState.java
@@ -21,11 +21,15 @@ import org.gradle.api.artifacts.Dependency;
 import org.gradle.api.artifacts.ModuleVersionIdentifier;
 import org.gradle.api.artifacts.component.ComponentIdentifier;
 import org.gradle.api.artifacts.component.ModuleComponentIdentifier;
+import org.gradle.api.artifacts.result.ResolvedVariantResult;
 import org.gradle.api.capabilities.CapabilitiesMetadata;
 import org.gradle.api.capabilities.Capability;
 import org.gradle.api.internal.artifacts.ivyservice.resolveengine.artifact.ArtifactSet;
+import org.gradle.api.internal.artifacts.result.DefaultResolvedVariantResult;
+import org.gradle.api.internal.attributes.AttributeDesugaring;
 import org.gradle.api.internal.attributes.ImmutableAttributes;
 import org.gradle.api.internal.component.ArtifactType;
+import org.gradle.internal.Describables;
 import org.gradle.internal.component.external.model.DefaultImmutableCapability;
 import org.gradle.internal.resolve.resolver.ArtifactResolver;
 import org.gradle.internal.resolve.resolver.ArtifactSelector;
@@ -36,14 +40,20 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
 public abstract class AbstractComponentGraphResolveState<T extends ComponentGraphResolveMetadata, S extends ComponentResolveMetadata> implements ComponentGraphResolveState, ComponentArtifactResolveState {
     private final T graphMetadata;
     private final S artifactMetadata;
+    private final AttributeDesugaring attributeDesugaring;
+    // The public view for each variant of this component
+    private final ConcurrentMap<VariantGraphResolveMetadata, ResolvedVariantResult> publicViews = new ConcurrentHashMap<>();
 
-    public AbstractComponentGraphResolveState(T graphMetadata, S artifactMetadata) {
+    public AbstractComponentGraphResolveState(T graphMetadata, S artifactMetadata, AttributeDesugaring attributeDesugaring) {
         this.graphMetadata = graphMetadata;
         this.artifactMetadata = artifactMetadata;
+        this.attributeDesugaring = attributeDesugaring;
     }
 
     @Override
@@ -98,6 +108,26 @@ public abstract class AbstractComponentGraphResolveState<T extends ComponentGrap
                 return builder.build();
             }
         };
+    }
+
+    @Override
+    public ResolvedVariantResult getVariantResult(VariantGraphResolveMetadata metadata, @Nullable ResolvedVariantResult externalVariant) {
+        if (externalVariant != null) {
+            // Don't cache the result
+            // Note that the external variant is a function of the metadata of the component, so should be constructed by this state object and cached rather than passed in
+            return createVariantResult(metadata, externalVariant);
+        } else {
+            return publicViews.computeIfAbsent(metadata, m -> createVariantResult(metadata, null));
+        }
+    }
+
+    private DefaultResolvedVariantResult createVariantResult(VariantGraphResolveMetadata metadata, @Nullable ResolvedVariantResult externalVariant) {
+        return new DefaultResolvedVariantResult(
+            getId(),
+            Describables.of(metadata.getName()),
+            attributeDesugaring.desugar(metadata.getAttributes()),
+            capabilitiesFor(metadata.getCapabilities()),
+            externalVariant);
     }
 
     @Nullable
